feat(editor): add useKeyboardShortcuts hook
