+++ apps/web/src/app/features/editor/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useEditorContent';
 export * from './usePageLeaveBlocker';
 export * from './useSaveFocusRestoration';
+export * from './useKeyboardShortcuts';
